convert to ts.
